requirements.txt
aniso8601==9.0.1
anyio==2.2.0
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==3.3.0
boto3==1.18.11
botocore==1.21.11
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.2
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.4.2
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.0.26
mysqlclient==2.0.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.3
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.2.0
pipreqs==0.4.10
platformdirs==2.3.0
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.17.3
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.1
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2021.1
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.25.1
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.3.20
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==2.0.1
yarg==0.1.9
